requirements.txt
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
Django==1.11.12
django-background-tasks==1.1.13
django-bootstrap4==0.0.6
django-compat==1.0.15
django-extensions==2.0.7
enum34==1.1.6
firebase-admin==2.9.1
futures==3.2.0
google-api-core==1.1.0
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.8.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
graphviz==0.8.3
grpcio==1.10.1
idna==2.6
lml==0.0.1
numpy==1.14.2
pandas==0.22.0
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydotplus==2.0.2
pyexcel-io==0.5.6
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
texttable==1.2.1
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.22
